Move custom sidebar selection helpers out of ContentView
